# aspnetcore-server-generated/src/IO.Swagger/Controllers/TargetModesApi.cs
/*
 * Service-Oriented Vehicle Diagnostic API
 *
 * Definition of ASAM SOVD API
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TargetModesApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns all available modes defined for the given entity.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Retrieve list of all supported modes of an entity</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/modes")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdModesGet")]
        public virtual IActionResult EntityCollectionEntityIdModesGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>This method returns all the details on the specified mode of the given entity</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="modeId">The identifier of a mode of an entity.</param>
        /// <param name="includeSchema">Specifies whether the response should include schema information or not. Default: false</param>
        /// <response code="200">Response: Get details of a single mode of an entity</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpGet]
        [Route("/{entity-collection}/{entity-id}/modes/{mode-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdModesModeIdGet")]
        public virtual IActionResult EntityCollectionEntityIdModesModeIdGet([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object modeId, [FromQuery]Object includeSchema)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <remarks>The method allows to control the state of the specified mode of the given entity explicitly.</remarks>
        /// <param name="entityCollection">Path to an entity collection</param>
        /// <param name="entityId">ID of an entity</param>
        /// <param name="modeId">The identifier of a mode of an entity.</param>
        /// <response code="200">Response: Get details of a single mode of an entity</response>
        /// <response code="500">An unexpected request occurred</response>
        /// <response code="0">An unexpected request occurred</response>
        [HttpPut]
        [Route("/{entity-collection}/{entity-id}/modes/{mode-id}")]
        [ValidateModelState]
        [SwaggerOperation("EntityCollectionEntityIdModesModeIdPut")]
        public virtual IActionResult EntityCollectionEntityIdModesModeIdPut([FromRoute][Required]Object entityCollection, [FromRoute][Required]Object entityId, [FromRoute][Required]Object modeId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            return StatusCode(200);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);


            throw new NotImplementedException();
        }
    }
}