# aspnetcore-server-generated/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Definition of ASAM SOVD API",
    "version" : "1.0.0",
    "title" : "Service-Oriented Vehicle Diagnostic API",
    "contact" : {
      "url" : "https://www.asam.net/project-detail/sovd-service-oriented-vehicle-diagnostics/"
    }
  },
...
  "paths" : {
...
    "/{entity-collection}/{entity-id}/modes" : {
      "get" : {
        "tags" : [ "target-modes" ],
        "description" : "This method returns all available modes defined for the given entity.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Retrieve list of all supported modes of an entity"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
    "/{entity-collection}/{entity-id}/modes/{mode-id}" : {
      "get" : {
        "tags" : [ "target-modes" ],
        "description" : "This method returns all the details on the specified mode of the given entity",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "mode-id",
          "in" : "path",
          "description" : "The identifier of a mode of an entity.",
          "required" : true
        }, {
          "name" : "include-schema",
          "in" : "query",
          "description" : "Specifies whether the response should include schema information or not. Default: false",
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Get details of a single mode of an entity"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      },
      "put" : {
        "tags" : [ "target-modes" ],
        "description" : "The method allows to control the state of the specified mode of the given entity explicitly.",
        "parameters" : [ {
          "name" : "entity-collection",
          "in" : "path",
          "description" : "Path to an entity collection",
          "required" : true
        }, {
          "name" : "entity-id",
          "in" : "path",
          "description" : "ID of an entity",
          "required" : true
        }, {
          "name" : "mode-id",
          "in" : "path",
          "description" : "The identifier of a mode of an entity.",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Response: Get details of a single mode of an entity"
          },
          "500" : {
            "description" : "An unexpected request occurred"
          },
          "default" : {
            "description" : "An unexpected request occurred"
          }
        }
      }
    },
...
  },
...
}